pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = "CRM - SEVE Finance"
authors = ["Van Tien NGUYEN <tiennguyenhust@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
gunicorn = "^23.0.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.6.0"
django-polymorphic = "^3.1.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.156.0"
celery = "^5.4.0"
redis = "^5.2.1"
pytz = "^2024.2"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
pre-commit = "^4.0.1"
isort = "^5.13.2"
black = "^24.10.0"
ipython = "^8.31.0"
factory-boy = "^3.3.1"
pyright = "^1.1.391"
autoflake = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
